+++ lib/conjur/api/router/v5.rb
@@ -157,7 +157,11 @@ module Conjur
         def ldap_sync_policy(credentials, config_name)
           RestClient::Resource.new(Conjur.configuration.core_url, credentials)['ldap-sync']["policy?config_name=#{fully_escape(config_name)}"]
         end
-        
+
+        def audit_show credentials
+          RestClient::Resource.new(Conjur.configuration.core_url, credentials)['audit']
+        end
+
         private
 
         def resource_annotations resource